backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
python-dotenv==1.0.0
pytest==7.4.0

